Support for local development endpoints
