+++ WordpressSync/debug.js
@@ -2,6 +2,8 @@
 const { Values } = require('../local.settings.json');
 Object.keys(Values).forEach(x=>process.env[x]=Values[x]); //Load local settings file for testing
 
+process.env.debug = true;
+
 //run the indexpage async
 const indexCode = require('./index');
 (async () => {
